refactor(backend): rename `user_prefs` module to `users`
