+++ apps/backend/src/main.rs
@@ -73,7 +73,7 @@ mod miscellaneous;
 mod models;
 mod providers;
 mod traits;
-mod user_prefs;
+mod users;
 mod utils;
 
 pub static VERSION: &str = env!("CARGO_PKG_VERSION");
